Full re-write with an internally-managed query parser.
This includes a re-implementation of our query execution into "Unit-of-work" "Statements" which serve as the glue between an IO operation for the define uow and the repository.

We've also dropped sqlite for the time-being as we focus on fleshing out the core feature-set and high-level API. Space has been left to implement new drivers for different dialects in the future.